moved code from circuit to internal call visitor
